requirements.txt
Faker==13.13.0
influxdb_client==1.32.0
paho_mqtt==1.6.1
